Rename test namespaces
