+++ composer.json
@@ -29,7 +29,7 @@
     },
     "autoload-dev":{
       "psr-4": {
-        "ArUtil\\Tests\\": "tests/"
+        "ArUtil\\Tests\\": "tests/ArUtil"
       }
     },
     "require-dev": {
