+++ tests/ArUtil/I18N/SoundexTest.php
@@ -2,8 +2,8 @@
 
 namespace ArUtil\Tests\Arabic;
 
-use ArUtil\Arabic;
-use ArUtil\Soundex;
+use ArUtil\I18N\Arabic;
+use ArUtil\I18N\Soundex;
 use ArUtil\Tests\AbstractTestCase;
 
 class SoundexTest extends AbstractTestCase
